Split validators between `l-plugins-api` and `l-plugins`
Plugin annotations are employed by `log4j-docgen` to generate
documentation.

Since the tool is compiled using Java 8 bytecode, we need to split
`log4j-plugins` into:

 * an API part that contains mostly annotations and interfaces,
 * an implementation part that can use all the improvements of Java 17.

In this commit:

 * we move the constraint annotation into a new artifact,
 * we make the `value` property of the `@Constraint` annotation
   optional,
 * we introduce a `ConstraintValidatorFactory` that produces the
   appropriate `ConstraintValidator` for each annotation. This way
   we don't need to reference `ConstraintValidator` implementations in
   the constraint annotations.
