+++ log4j-plugins-api/src/main/java/org/apache/logging/log4j/plugins/PluginException.java
