+++ log4j-plugins-api/src/main/java/org/apache/logging/log4j/plugins/validation/constraints/RequiredClass.java
@@ -1,0 +1,45 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to you under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.logging.log4j.plugins.validation.constraints;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.apache.logging.log4j.plugins.validation.Constraint;
+
+/**
+ * Marks a plugin builder field or plugin factory parameter as required.
+ *
+ * @since 3.0.0
+ */
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE})
+@Constraint
+public @interface RequiredClass {
+
+    /** The fully qualified name of the Class that is required */
+    String value();
+
+    /**
+     * The message to be logged if this constraint is violated. This should normally be overridden.
+     * @return the message to be logged if the constraint is violated.
+     */
+    String message() default "The required class is not present";
+}
