+++ log4j-plugins-api/src/main/java/org/apache/logging/log4j/plugins/validation/spi/package-info.java
@@ -14,15 +14,9 @@
  * See the license for the specific language governing permissions and
  * limitations under the license.
  */
-
-/**
- * Validation annotations.
- *
- * @since 2.1
- */
 @Export
-@Version("1.0.0")
-package org.apache.logging.log4j.plugins.validation.constraints;
+@Version("3.0.0")
+package org.apache.logging.log4j.plugins.validation.spi;
 
 import org.osgi.annotation.bundle.Export;
 import org.osgi.annotation.versioning.Version;
