+++ log4j-plugins/src/main/java/org/apache/logging/log4j/plugins/internal/validation/RequiredPropertyValidator.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.logging.log4j.plugins.validation.validators;
+package org.apache.logging.log4j.plugins.internal.validation;
 
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.plugins.validation.ConstraintValidator;
