+++ log4j-plugins/src/main/java/org/apache/logging/log4j/plugins/internal/validation/ValidHostValidator.java
@@ -14,36 +14,48 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.logging.log4j.plugins.validation.validators;
+package org.apache.logging.log4j.plugins.internal.validation;
 
+import java.net.InetAddress;
+import java.net.UnknownHostException;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.plugins.validation.ConstraintValidator;
-import org.apache.logging.log4j.plugins.validation.constraints.RequiredClass;
+import org.apache.logging.log4j.plugins.validation.constraints.ValidHost;
 import org.apache.logging.log4j.status.StatusLogger;
-import org.apache.logging.log4j.util.LoaderUtil;
 
 /**
- * Validator that checks for the existence of a class that is required.
+ * Validator that checks an object to verify it is a valid hostname or IP address. Validation rules follow the same
+ * logic as in {@link InetAddress#getByName(String)}.
  *
- * @since 3.0.0
+ * @since 2.8
  */
-public class RequiredClassValidator implements ConstraintValidator<RequiredClass> {
+public class ValidHostValidator implements ConstraintValidator<ValidHost> {
 
     private static final Logger LOGGER = StatusLogger.getLogger();
 
-    private RequiredClass annotation;
+    private ValidHost annotation;
 
     @Override
-    public void initialize(final RequiredClass anAnnotation) {
-        this.annotation = anAnnotation;
+    public void initialize(final ValidHost annotation) {
+        this.annotation = annotation;
     }
 
     @Override
     public boolean isValid(final String name, final Object value) {
-        final boolean isValid = LoaderUtil.isClassAvailable(annotation.value());
-        if (!isValid) {
-            LOGGER.error("{} cannot be used. {} is unavailable.", name, annotation.value());
+        if (value == null) {
+            LOGGER.error(annotation.message());
+            return false;
         }
-        return isValid;
+        if (value instanceof InetAddress) {
+            // InetAddress factory methods all have built in validation
+            return true;
+        }
+        try {
+            InetAddress.getByName(value.toString());
+            return true;
+        } catch (final UnknownHostException e) {
+            LOGGER.error(annotation.message(), e);
+            return false;
+        }
     }
 }
