+++ pom.xml
@@ -263,6 +263,7 @@
     <module>log4j-perf-test</module>
     <module>log4j-plugin-processor</module>
     <module>log4j-plugins</module>
+    <module>log4j-plugins-api</module>
     <module>log4j-plugins-test</module>
     <module>log4j-script</module>
     <module>log4j-slf4j2-impl</module>
@@ -492,6 +493,12 @@
       <dependency>
         <groupId>org.apache.logging.log4j</groupId>
         <artifactId>log4j-mongodb4</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.apache.logging.log4j</groupId>
+        <artifactId>log4j-plugins-api</artifactId>
         <version>${project.version}</version>
       </dependency>
